Rest Expressions
Rests (...) can now be combined with arbitrary patterns instead of just
identifiers. This has the effect of applying the pattern to each item in
the list. Identifiers within a rest expression will pass on an array
containing all values for that pattern. So a pattern `[...{name}]` would
check that each item in the array is an object that has a name
attribute. It would then pass on an array to the function containing all
the names.

This doesn't really work for rests within objects, so rests have been
removed from object entirely. Object matching is only non-strict now.
There didn't seem to be much of a use for strict object matching, and
the added rest to perform non-strict matching was irritating. The
compiler logic was also complex, and its good to see it go.